ClusterCenter/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0

# 扩展功能依赖（可选）
paho-mqtt==1.6.1  # MQTT 支持
psycopg2-binary==2.9.9  # PostgreSQL 支持
sqlalchemy==2.0.23  # ORM（可选）

# 高级优化功能依赖（可选）
torch==2.1.0  # PyTorch（LSTM、Transformer 模型）
scikit-learn==1.3.2  # 数据预处理（可选）

# 分布式集群依赖（可选）
aiohttp==3.9.1  # 异步 HTTP 客户端（用于 RPC 通信）

# 服务发现依赖（可选）
# consul==1.1.0  # Consul Python 客户端（如需要）
# python-etcd3==0.1.5  # etcd Python 客户端（如需要）

# gRPC 支持（可选）
grpcio==1.60.0  # gRPC 核心库
grpcio-tools==1.60.0  # gRPC 工具（用于生成代码）